requirements.txt
annotated-types==0.6.0
appnope==0.1.3
asttokens==2.4.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.3
blis==0.7.11
bs4==0.0.1
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
cloudpathlib==0.16.0
comm==0.1.4
confection==0.1.3
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.0
decorator==5.1.1
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.0/en_core_web_md-3.7.0-py3-none-any.whl#sha256=d37c3d60738be409a04f83a51421075b94fc02efad99bb8e62ec9018b1f99cb0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
exceptiongroup==1.1.3
executing==2.0.1
Flask==3.0.0
Flask-Cors==4.0.0
fonttools==4.44.0
google-api-core==2.15.0
google-auth==2.26.1
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-images-download==2.8.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
gunicorn==21.2.0
h11==0.14.0
idna==3.4
ipykernel==6.26.0
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.5.0
jupyter_core==5.5.0
kiwisolver==1.4.5
langcodes==3.3.0
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
movieposters==0.0.7
murmurhash==1.0.10
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.26.1
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.2
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
platformdirs==3.11.0
preshed==3.0.9
prompt-toolkit==3.0.39
protobuf==4.25.2
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
selenium==4.15.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
thinc==8.2.1
threadpoolctl==3.2.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
trio==0.23.0
trio-websocket==0.11.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wasabi==1.1.2
wcwidth==0.2.9
weasel==0.3.3
Werkzeug==3.0.1
wsproto==1.2.0
